stockbot_discord/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
cycler==0.10.0
discord==1.0.1
discord.py==1.3.3
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
multidict==4.7.6
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.18.4
pyimgur==0.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
schedule==0.6.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
webdriver-manager==2.5.1
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2